Refactor strategy_form_inputs method to remove rubocop method length alert
